requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
bcrypt==5.0.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.124.0
firebase-admin==6.9.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.5.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.0.1
idna==3.11
jose==1.0.0
msgpack==1.1.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.23
pydantic==1.10.12
# pydantic_core removed — using pydantic v1 which doesn't need pydantic_core
gunicorn==20.1.0
psycopg2-binary==2.9.11
PyJWT==2.9.0
pyparsing==3.1.4
python-jose==3.4.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.44.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.33.0
